+++ src/interfaces/sim800-command.enums.ts
@@ -2,3 +2,11 @@ export enum CmgfMode {
   Pdu = '0',
   Text = '1',
 }
+
+export enum CmglStat {
+  ReceivedUnread = 0,
+  ReceivedRead = 1,
+  StoredUnsent = 2,
+  StoredSent = 3,
+  All = 4,
+}