+++ src/interfaces/sim800-incoming-sms.interface.ts
@@ -1,4 +1,5 @@
 export interface Sim800IncomingSms {
   number: string;
   text: string;
+  date: Date;
 }